rov_final/requirements.txt
opencv-python>=4.9
numpy>=1.24
pyserial>=3.5
matplotlib>=3.7
picamera2>=0.3
pyzmq>=25.0
msgpack>=1.0
pytest>=7.0
